Reworking the book creator to use service bus
