+++ AzureCodeCamp/PancakeProwler.BookCreator/IMailSender.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Linq;
+using System.Collections.Generic;
+using PancakeProwler.Data.Common.Models;
+
+namespace PancakeProwler.BookCreator
+{
+    internal interface IMailSender
+    {
+        void SendCreationEMail(BookCreationRequest request);
+    }
+}
