+++ AzureCodeCamp/PancakeProwler.BookCreator/packages.config
@@ -9,5 +9,6 @@
   <package id="Sendgrid" version="5.0.0" targetFramework="net45" />
   <package id="SendGrid.SmtpApi" version="1.1.3" targetFramework="net45" />
   <package id="System.Spatial" version="5.6.3" targetFramework="net45" />
+  <package id="WindowsAzure.ServiceBus" version="2.5.3.0" targetFramework="net45" />
   <package id="WindowsAzure.Storage" version="4.3.0" targetFramework="net45" />
 </packages>
